HW - LangChain II/solution/requirements-minimal.txt
langchain==0.1.16
langchain-core==0.1.43
langchain-community==0.0.33
langchain-openai==0.1.3
langchainhub==0.1.15
sentence-transformers==2.7.0
python-dotenv==1.0.1
pandas==2.2.2
faiss-cpu==1.8.0
fastapi==0.115.5
uvicorn==0.32.0
pytest==8.3.3
